@@ -1,18 +1,18 @@
 name: flutter_notif_tone_picker
 description: Plugin for pick notification tone uri (Android Only)
-version: 0.0.1
+version: 1.0.0
 author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_plugin_android_lifecycle: ^2.0.0
+  flutter_plugin_android_lifecycle: ^2.0.1
 
 dev_dependencies:
   flutter_test:
